requirements.txt
# Core
joblib==1.5.2
numpy==2.3.2
pandas==2.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2

# Machine Learning
scikit-learn==1.6.1
xgboost==2.1.2

# Azure
azure-storage-blob==12.23.1

# Environment Variables
python-dotenv==1.0.1


